# Backend/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: ReForest API
  description: API para la gestión de la plataforma ReForest, que permite a los usuarios realizar donaciones para la reforestación y seguimiento de proyectos.
  version: "0.0.1-SNAPSHOT"
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
...
paths:
...
  /especies:
    get:
      tags:
        - Especies
      summary: Obtener lista de todas las especies
      description: Devuelve una lista con todas las especies de árboles disponibles.
      operationId: getAllEspecies
      responses:
        '200':
          description: Lista de especies obtenida con éxito.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Especie'
              example:
                - id: 1
                  nombreComun: "Pino Carrasco"
                  precioPlantacion: 15.50
                  disponible: true
                - id: 2
                  nombreComun: "Encina"
                  precioPlantacion: 25.00
                  disponible: false
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
        - Especies
      summary: Crear una nueva especie
      description: Registra una nueva especie de árbol en el sistema.
      operationId: createEspecie
      requestBody:
        description: Datos de la nueva especie a crear.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EspecieInput'
      responses:
        '201':
          description: Especie creada con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Especie'
              example:
                id: 3
                nombreComun: "Roble"
                descripcion: "Árbol robusto de hoja caduca."
                precioPlantacion: 30.00
                zonaGeografica: "Clima Atlántico"
                co2AnualKg: 25.0
                alturaMaximaM: 40
                disponible: true
                fechaTemporada: "2026-10-01"
                imageUrl: "http://example.com/images/roble.jpg"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /especies/{id}:
    get:
      tags:
        - Especies
      summary: Obtener una especie por su ID
      description: Devuelve los detalles de una especie específica.
      operationId: getEspecieById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Especie encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Especie'
              example:
                id: 1
                nombreComun: "Pino Carrasco"
                descripcion: "Conífera resistente a la sequía."
                precioPlantacion: 15.50
                zonaGeografica: "Mediterráneo"
                co2AnualKg: 22.5
                alturaMaximaM: 20
                disponible: true
                fechaTemporada: "2026-09-15"
                imageUrl: "http://example.com/images/pino.jpg"
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Especies
      summary: Actualizar una especie existente
      operationId: updateEspecie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EspecieInput'
      responses:
        '200':
          description: Especie actualizada con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Especie'
              example:
                id: 1
                nombreComun: "Pino Carrasco (Actualizado)"
                precioPlantacion: 16.00
                disponible: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Especies
      summary: Desactivar una especie (borrado lógico)
      description: Realiza un borrado lógico de la especie, marcándola como 'no disponible'. La especie no se elimina de la base de datos y puede ser recuperada. Ya no aparecerá en las listas por defecto.
      operationId: deleteEspecie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Especie desactivada con éxito.
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  schemas:
...
    Especie:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        nombreComun:
          type: string
        descripcion:
          type: string
        precioPlantacion:
          type: number
          format: double
        zonaGeografica:
          type: string
        co2AnualKg:
          type: number
          format: double
        alturaMaximaM:
          type: integer
        disponible:
          type: boolean
        fechaTemporada:
          type: string
          format: date
        imageUrl:
          type: string
          format: uri
    EspecieInput:
      type: object
      properties:
        nombreComun:
          type: string
        descripcion:
          type: string
        precioPlantacion:
          type: number
          format: double
        zonaGeografica:
          type: string
        co2AnualKg:
          type: number
          format: double
        alturaMaximaM:
          type: integer
        disponible:
          type: boolean
        fechaTemporada:
          type: string
          format: date
        imageUrl:
          type: string
          format: uri
      required:
        - nombreComun
        - precioPlantacion
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
          description: Un código de error numérico HTTP.
          example: 404
        message:
          type: string
          description: Una descripción del error legible por humanos.
          example: "El recurso solicitado no fue encontrado."
  responses:
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 404
            message: "El recurso solicitado no se ha encontrado."
    BadRequest:
      description: Solicitud incorrecta
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            message: "Los datos de la solicitud son inválidos."
...
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            message: "Ha ocurrido un error inesperado en el servidor."